Put Sync tests in their own test fiels
